extract server thread
