+++ src/lib.rs
@@ -5,3 +5,6 @@ mod app;
 pub use crate::app::parse_arg;
 
 pub mod files;
+
+mod server;
+pub use crate::server::spawn_server;
