+++ src/main.rs
@@ -4,12 +4,11 @@ extern crate notify;
 extern crate ws;
 extern crate preview_image_folder;
 
-use preview_image_folder::{genpage, files};
+use preview_image_folder::{files, spawn_server};
 use std::sync::mpsc::channel;
 use std::thread;
 use std::thread::sleep;
 use std::time::Duration;
-use urlencoding::decode;
 
 // v5
 // use notify::{Watcher, RecursiveMode, Event};
@@ -17,73 +16,7 @@ use urlencoding::decode;
 // v4
 use notify::DebouncedEvent::{Create, Remove, Rename, Write};
 use notify::{watcher, RecursiveMode, Watcher};
-use ws::{connect, listen, CloseCode, Handler, Message, Request, Response, Sender};
-
-fn response_with_contenttype<R>(
-    status: u16,
-    reason: R,
-    body: Vec<u8>,
-    contenttype: Vec<u8>,
-) -> Response
-where
-    R: Into<String>,
-{
-    let mut res = Response::new(status, reason, body);
-    res.headers_mut().push(("Content-type".into(), contenttype));
-    res
-}
-
-// Server web application handler
-struct Server<'a> {
-    out: Sender,
-    target: &'a str,
-}
-
-impl Handler for Server<'_> {
-    //
-    fn on_request(&mut self, req: &Request) -> ws::Result<Response> {
-        // Using multiple handlers is better (see router example)
-        match req.resource() {
-            // The default trait implementation
-            "/ws" => Response::from_request(req),
-
-            // Create a custom response
-            "/" => Ok(response_with_contenttype(
-                200,
-                "OK",
-                genpage(),
-                "text/html; charset=UTF-8".into(),
-            )),
-
-            "/index.html" => Ok(response_with_contenttype(
-                200,
-                "OK",
-                genpage(),
-                "text/html; charset=UTF-8".into(),
-            )),
-
-            "/images.json" => Ok(response_with_contenttype(
-                200,
-                "OK",
-                files::list_images(self.target),
-                "application/json; charset=UTF-8".into(),
-            )),
-
-            other => Ok(Response::new(
-                200,
-                "OK",
-                files::read_file(format!("{}{}", self.target, decode(other).expect("UTF-8")).as_str()),
-            )),
-            // _ => Ok(Response::new(404, "Not Found", b"404 - Not Found".to_vec())),
-        }
-    }
-
-    // Handle messages recieved in the websocket (in this case, only on /ws)
-    fn on_message(&mut self, msg: Message) -> ws::Result<()> {
-        // Broadcast to all connections
-        self.out.broadcast(msg)
-    }
-}
+use ws::{connect, CloseCode};
 
 fn main() {
     // Setup logging
@@ -109,13 +42,7 @@ fn main() {
     // Server thread
     // Listen on an address and call the closure for each connection
     let server_target = format!("{}", target);
-    let server = thread::spawn(move || {
-        listen(server_url, |out| Server {
-            out,
-            target: &server_target,
-        })
-        .unwrap()
-    });
+    let server = spawn_server(server_url, server_target);
 
     // Give the server a little time to get going
     sleep(Duration::from_millis(10));
